Fix not running callback in pipeline custom command
The issue is that Pipeline.prototype.sendCommand(...) returns
this (a Pipeline instance for chaining) instead of the command's
Promise (which every other sendCommand method seems to return). When
this Pipeline instance filters down to line 24 in lib/script.js
Script.prototype.execute(...), the condition (result instanceof
Promise) fails and any callback the user might have passed in never
gets attached to the Promise, and is subsequently never run.